combine AppTypeCommand into ContextualCommand to simplify
