+++ lib/shopify-cli/commands/serve/node.rb
@@ -4,6 +4,12 @@ module ShopifyCli
   module Commands
     class Serve
       class ServeNode < ShopifyCli::Commands::Serve
+        options do |parser, flags|
+          parser.on('--host=HOST') do |h|
+            flags[:host] = h.gsub('"', '')
+          end
+        end
+
         def call(*)
           setup
           CLI::UI::Frame.open('Running server...') do
