+++ lib/shopify_cli.rb
@@ -95,7 +95,6 @@ module ShopifyCli
   end
 
   autoload :API, 'shopify-cli/api'
-  autoload :AppTypeCommand, 'shopify-cli/app_type_command'
   autoload :AppTypeRegistry, 'shopify-cli/app_type_registry'
   autoload :AppTypes, 'shopify-cli/app_types'
   autoload :Command, 'shopify-cli/command'
